Move HlslKnownKeywords to shared generation project
